+++ examples/framework-examples/vue-example/vue3-cli-applet/src/main.js
@@ -1,12 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import sos from '@signageos/front-applet'
 
-Vue.config.productionTip = false
-
 sos.onReady().then(() => {
   console.log('Ready to rock')
-  new Vue({
-    render: h => h(App),
-  }).$mount('#app')
+  createApp(App).mount('#app')
 })